ViT_explain/requirements.txt
pytorch-transformers==1.1.0
torch
seaborn==0.11.2
matplotlib
PIL==8.2.0
torchvision==0.11.0a0
numpy